Moved baselib.datastore in commonlib
