+++ openquake/calculators/tests/event_based_test.py
@@ -26,9 +26,9 @@ import numpy.testing
 
 from openquake.baselib.hdf5 import read_csv
 from openquake.baselib.general import countby, gettemp
-from openquake.baselib.datastore import read
 from openquake.hazardlib import nrml, InvalidFile
 from openquake.hazardlib.sourceconverter import RuptureConverter
+from openquake.commonlib.datastore import read
 from openquake.commonlib.writers import write_csv
 from openquake.commonlib.util import max_rel_diff_index
 from openquake.commonlib.calc import gmvs_to_poes
