+++ openquake/commands/purge.py
@@ -18,8 +18,7 @@
 import os
 import re
 import getpass
-from openquake.baselib import datastore
-from openquake.commonlib.logs import dbcmd
+from openquake.commonlib import logs, datastore
 
 datadir = datastore.get_datadir()
 
@@ -28,7 +27,7 @@ def purge_one(calc_id, user, force):
     """
     Remove one calculation ID from the database and remove its datastore
     """
-    dbcmd('del_calc', calc_id, user, force)
+    logs.dbcmd('del_calc', calc_id, user, force)
     f1 = os.path.join(datadir, 'calc_%s.hdf5' % calc_id)
     f2 = os.path.join(datadir, 'calc_%s_tmp.hdf5' % calc_id)
     for f in [f1, f2]:
