+++ openquake/commands/reaggregate.py
@@ -19,8 +19,8 @@
 import os
 import getpass
 import logging
-from openquake.baselib import parallel, datastore
-from openquake.commonlib import logs
+from openquake.baselib import parallel
+from openquake.commonlib import logs, datastore
 from openquake.calculators.post_risk import PostRiskCalculator
 from openquake.engine import engine
