+++ openquake/commands/show_attrs.py
@@ -15,7 +15,7 @@
 #
 # You should have received a copy of the GNU Affero General Public License
 # along with OpenQuake. If not, see <http://www.gnu.org/licenses/>.
-from openquake.baselib import datastore
+from openquake.commonlib import datastore
 import h5py
