+++ openquake/commonlib/tests/hazard_writers_test.py
@@ -17,7 +17,6 @@
 # along with OpenQuake. If not, see <http://www.gnu.org/licenses/>.
 
 import sys
-import json
 import numpy
 import os
 import io
