+++ openquake/commonlib/tests/logictree_test.py
@@ -15,6 +15,7 @@
 #
 # You should have received a copy of the GNU Affero General Public License
 # along with OpenQuake. If not, see <http://www.gnu.org/licenses/>.
+
 import os
 import codecs
 import unittest
