+++ openquake/engine/tools/make_html_report.py
@@ -23,9 +23,8 @@ from datetime import date, datetime, timedelta
 import itertools
 from docutils.examples import html_parts
 
-from openquake.baselib.datastore import read
 from openquake.calculators.views import view_fullreport
-from openquake.commonlib.logs import dbcmd
+from openquake.commonlib import logs, datastore
 
 tablecounter = itertools.count(0)
 
@@ -164,17 +163,17 @@ def make_report(isodate='today'):
     tag_contents = []
 
     # the fetcher returns an header which is stripped with [1:]
-    jobs = dbcmd(
+    jobs = logs.dbcmd(
         'fetch', ALL_JOBS, isodate.isoformat(), isodate1.isoformat())
     page = '<h2>%d job(s) finished before midnight of %s</h2>' % (
         len(jobs), isodate)
     for job_id, user, status, ds_calc in jobs:
         tag_ids.append(job_id)
         tag_status.append(status)
-        [stats] = dbcmd('fetch', JOB_STATS, job_id)
+        [stats] = logs.dbcmd('fetch', JOB_STATS, job_id)
         (job_id, user, start_time, stop_time, status, duration) = stats
         try:
-            ds = read(job_id, datadir=os.path.dirname(ds_calc))
+            ds = datastore.read(job_id, datadir=os.path.dirname(ds_calc))
             txt = view_fullreport('fullreport', ds)
             report = html_parts(txt)
         except Exception as exc:
