+++ openquake/server/db/actions.py
@@ -22,7 +22,8 @@ import operator
 from datetime import datetime
 
 from openquake.hazardlib import valid
-from openquake.baselib import datastore, general
+from openquake.baselib import general
+from openquake.commonlib import datastore
 from openquake.calculators.export import export
 from openquake.server import __file__ as server_path
 from openquake.server.db.schema.upgrades import upgrader
